feat: :sparkles: Document services
Implementation of doc and template services, now we can create a new template, a new document using a template, and get my documents

Disclaimer: Should have implemented the template services in a template feature, but I didn´t think that it was required to have a template before creating a document. Oops!